Detect dealing item cards to player
Item cards can be purchased by players or dealt to players when they
move to an item location. The `in_hand` tag is used to distinguish
whether a card is available to be purchased by players. Today only
purchased item cards are tagged with `in_hand`, meaning dealt cards are
still considered available for purchase, which is incorrect.

This change tags item cards with `in_hand` when they are dealt to
players.