Implement custom counter
The TTS built-in counter lacks one critical utility -- an event handler
like onClick. In many scenarios we needed to do things as a result of
counter being incremented or decremented. One example being when
happiness of a player changes, we need to recalculate VP. We achieve
this by creating a custom counter build on top of TTS buiilt-in custom
tile, with 3 buttons generated on load: one for increment, one for
decrement, one for displaying counter value.