Implement happiness counter from custom counter
Happiness counter is built on top of the custom counter for happiness
related victory points calculations. A happiness counter's starting
value is 3, and is always associated with a victory point counter. When
happiness is incremented to or decremented from multiples of 5, victory
points is incremented or decremented, respectively. Negative hapiness
does not affect victory points.