Fix custom button drawing multiple cards issue
Before I discoved the use of scriptiong zone we relied on the use of a
tag called "in_hand" for detecting whether a card can be drawn by a
player or not.  This method is not reliable because we have to update
the tag whenever the card enters the Hand zone, and sometimes during
development the updating mechanism is not fully finished, leaving the
board in an inconsistent state, causing hard to reason bugs.

One such bug surfaced with the addition of draw card buttons. The card
drawn from the deck is first compared with a card name prefix, and then
checked for the in_hand tag, but because the board was in an
inconsistent state, some cards in deck already have the tag. Those cards
would be left in the middle of the board, until one card with the
prefix, and without the tag is drawn.

This change removes the dependency on the in_hand tag, and relies
exclusively on whether the card is in its own deck zone to determine if
it can be drawn again.