Fix issues when deck is completed drawn
When the second to last card within a deck is taken out, the deck object
will be destroyed, and the last card in a deck becomes a card object.
Deck and card are two different objects with different built in
functions, function such as getAllObjects only works with a deck, not a
card, enhance fucntion deck.drawFromDeck to deal with such scenarios.