Auto return item and fate cards to deck
When an item or fate card is played, we relied on game master to
manually put those cards back into the decks, this is a lot of
unnecessary mechanical work on game master.

With this change, when any item or fate cards is placed on the game
board, it is automatically returned to its deck.