Hide item cards when enter hand zone
Game rules allow item cards to be stolen from one player to another,
however as soon as item card leaves hand zone it is revealed to every
player, which is undesirable.

With this change, as soon as an item card is drawn to a player's hand,
it is hidden from everyone and except the holding player and game
master. The card remains hidden until played on board.

A 3 second delay is added before a played card is returned to its deck,
to allow playered to read the content of the card with ample time.