Reveals item cards when it's played on the board
If players wants to see the item card and don't want it returned to the
deck yet, they can place the item card anywhere on the table besides the
main board.

With this change we can also reduce the time delay for returning cards
to deck, since players can keep the card on table for viewing.