getDeckFromZoneGUID only returns deck or card GUID
Players may be tossing random objects around and end up in one of the
deck zones, which assumes only one deck exists in the zone. In this case
getDeckFromZoneGUID will return the GUID of that random object instead
of the deck or card in zone. This change adds safety check to ensure
only deck or card's GUID is returned.